Fixed compiler error

+++ CruelerThanDAT/inc/FileNodes.h
@@ -2792,7 +2792,7 @@ public:
 
 
 					if (type == 0x4) {
-						BnkEventObject evnObj{ type, size, uid };
+						BnkEventObject evnObj{ static_cast<uint8_t>(type), size, uid };
 						int eventCounts = reader.ReadUINT32();
 						for (int i = 0; i < eventCounts; i++) {
 							evnObj.ids.push_back(reader.ReadUINT32());
